+++ test/callApi.test.js
@@ -11,13 +11,13 @@ describe('testing api', () => {
   it('calls GibliApi with the rigth url', () => {
     fetch.mockResponseOnce(JSON.stringify(filmsApi));
 
-    //assert on the response
+    // assert on the response
     callApi(ENDPOINTGOOD);
 
     expect(fetch).toHaveBeenCalled();
     expect(fetch).toHaveBeenCalledWith(ENDPOINTGOOD);
 
-    //assert on the times called and arguments given to fetch
+    // assert on the times called and arguments given to fetch
     // expect(fetch.mock.calls.length).toEqual(1);
     // expect(fetch.mock.calls[0][0]).toEqual(ENDPOINTGOOD);
   });
@@ -26,7 +26,7 @@ describe('testing api', () => {
     const error = new Error('errorrrrrrrrrrrr');
     fetch.mockReject(error);
 
-    //assert on the response
+    // assert on the response
     const result = await callApi(ENDPOINTGOOD);
 
     expect(result).toEqual(error);
@@ -35,7 +35,7 @@ describe('testing api', () => {
   it('calls GibliApi and returns data to me', async () => {
     fetch.mockResponseOnce(JSON.stringify(filmsApi));
 
-    //assert on the response
+    // assert on the response
     // callApi(ENDPOINTGOOD).then((res) => {
     //   expect(res).toEqual(filmsApi);
     // });
