+++ test/fixtures/variables-fixtures.js
@@ -1,4 +1,22 @@
 const ENDPOINTGOOD = './fixtures.filmsApi.js';
 const ENDPOINTBAD = 'https://ghibliapi.herokuappjasdjkh';
 
-export { ENDPOINTGOOD, ENDPOINTBAD };
+const mockFilms = [
+  {
+    title: 'hola',
+    description: 'no',
+    director: 'patricia',
+  },
+  {
+    title: 'adios',
+    description: 'no',
+    director: 'chus',
+  },
+  {
+    title: 'byeo',
+    description: 'yes',
+    director: 'luis',
+  },
+];
+
+export { ENDPOINTGOOD, ENDPOINTBAD, mockFilms };
